requirements.txt
asgiref==3.7.2
awscli==1.29.28
botocore==1.31.28
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.4
Django==4.2.4
djangorestframework==3.14.0
docutils==0.16
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
jmespath==1.0.1
packaging==23.1
psycopg2-binary==2.9.7
pyasn1==0.5.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rsa==4.7.2
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
